pyproject.toml
[project]
name = "topgridscholar"
version = "0.1.0"
description = "Academic paper search & download tool — IEEE · Nature · Semantic Scholar"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "streamlit>=1.30.0",
    "playwright>=1.40.0",
    "pandas>=2.0.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
topgridscholar = "topgridscholar.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
